feat: add the option for generating mock results
