+++ apps/react-ui/client/tsconfig.json
@@ -23,7 +23,9 @@
 			"@providers/*": ["./src/providers/*"],
 			"@components/*": ["./src/components/*"],
 			"@pages/*": ["./src/pages/*"],
-			"@styles/*": ["./src/styles/*"]
+			"@styles/*": ["./src/styles/*"],
+			"@tests/*": ["./src/tests/*"],
+			"@utils/*": ["./src/utils/*"]
 		}
 	},
 	"include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
